Move to Matlab
I’ve decided to progress with Matlab with the data analysis. I am more
familiar with he program which will allow me to work quicker. I will
continue to scrape data with Python, but that is as far as I intend to
use Python.

Project organization will be conducted in OneNote, thus the removal of
the readme file.

I have downloaded a couple more spread sheets and imported them into
Matlab. I created a variable spreadsheet that I will edit, rather than
in the program. Next steps are to better organize and document the
process.