Entered First Lineups
This update marks the beginning of implementing this script on the
Fanduel website. Lineup optimization is greatly improved to reduce the
time it takes to create lineups to less than 5 minutes. Improvements to
rating system, such as home and away games and multiplier, are
implemented in this update. From here on out, tweaking will be required
to improve results. New script for lineup export is in progress.